requirements.txt
tensorflow==2.12.0
keras==2.12.0
Flask==2.2.3
Pillow==9.5.0
numpy==1.23.5
werkzeug==2.2.3
gunicorn==20.1.0